# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Broker API",
    "version": "1.0.0",
    "description": "This is the APIs documentation for Broker Platform"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Development server" },
    { "url": "https://broker-six-navy.vercel.app/", "description": "Production server" }
  ],
...
  "paths": {
...
    "/api/user": {
      "get": {
        "tags": ["User"],
  "summary": "List users",
  "description": "Requires SUPER_ADMIN, ADMIN, or AGENT role. Agents only see clients they created.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserListResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized - missing or invalid token" },
          "403": { "description": "Forbidden - insufficient role" },
          "500": { "description": "Server error" }
        }
      },
      "post": {
        "tags": ["User"],
  "summary": "Create a new user",
  "description": "Requires SUPER_ADMIN, ADMIN, or AGENT role. Agents can only create client accounts. Accepts full signup payload plus optional administrative fields.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUserRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "400": { "description": "Invalid data" },
          "401": { "description": "Unauthorized - missing or invalid token" },
          "403": { "description": "Forbidden - insufficient role" },
          "409": { "description": "Conflict - user with provided unique field already exists" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/api/user/{id}": {
      "get": {
        "tags": ["User"],
  "summary": "Get user by ID",
  "description": "Requires SUPER_ADMIN, ADMIN, or AGENT role. Agents can only access clients they created.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized - missing or invalid token" },
          "403": { "description": "Forbidden - insufficient role" },
          "404": { "description": "User not found" },
          "500": { "description": "Server error" }
        }
      },
      "patch": {
        "tags": ["User"],
  "summary": "Update user by ID",
  "description": "Requires SUPER_ADMIN, ADMIN, or AGENT role. Agents can only modify clients they created. Accepts partial updates.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateUserRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "400": { "description": "Invalid data" },
          "401": { "description": "Unauthorized - missing or invalid token" },
          "403": { "description": "Forbidden - insufficient role" },
          "404": { "description": "User not found" },
          "409": { "description": "Conflict - user with provided unique field already exists" },
          "500": { "description": "Server error" }
        }
      },
      "delete": {
        "tags": ["User"],
  "summary": "Delete user by ID",
  "description": "Requires SUPER_ADMIN, ADMIN, or AGENT role. Agents can only remove clients they created.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "User deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized - missing or invalid token" },
          "403": { "description": "Forbidden - insufficient role" },
          "404": { "description": "User not found" },
          "500": { "description": "Server error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SignupRequest": {
        "type": "object",
        "properties": {
          "firstName": { "type": "string", "minLength": 3 },
          "lastName": { "type": "string", "minLength": 3 },
          "email": { "type": "string", "format": "email" },
          "phoneCountryCode": { "type": "string", "pattern": "^\\+[0-9]{1,4}$" },
          "phone": {
            "type": "string",
            "pattern": "^[0-9]{4,15}$",
            "description": "Digits only; spaces are removed before validation."
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "description": "Must contain uppercase, lowercase, numeric, and special characters."
          },
          "confirmPassword": { "type": "string", "minLength": 8 },
          "idNumber": { "type": "string", "minLength": 1 },
          "csdNumber": { "type": "string", "description": "Optional Central Securities Depository number" },
          "passportPhoto": {
            "type": "string",
            "description": "Secure URL returned by the Cloudinary upload endpoint for the user's passport photo"
          },
          "idDocument": {
            "type": "string",
            "description": "Secure URL returned by the Cloudinary upload endpoint for the user's identification document"
          },
          "dateOfBirth": { "type": "string", "format": "date" },
          "gender": {
            "type": "string",
            "enum": ["male", "female"]
          },
          "country": { "type": "string", "minLength": 1 },
          "city": { "type": "string", "minLength": 1 },
          "occupation": { "type": "string", "minLength": 1 },
          "investmentExperience": { "type": "string", "minLength": 1 }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "phoneCountryCode",
          "phone",
          "password",
          "confirmPassword",
          "idNumber",
          "passportPhoto",
          "idDocument",
          "dateOfBirth",
          "gender",
          "country",
          "city",
          "occupation",
          "investmentExperience"
        ]
      },
      "UserRole": {
        "type": "string",
        "enum": ["SUPER_ADMIN", "ADMIN", "AGENT", "COMPANY", "CLIENT"]
      },
      "CreateUserRequest": {
        "allOf": [
          { "$ref": "#/components/schemas/SignupRequest" },
          {
            "type": "object",
            "properties": {
              "notificationPreferences": {
                "type": "object",
                "additionalProperties": true
              },
              "role": { "$ref": "#/components/schemas/UserRole" },
              "isVerified": { "type": "boolean" }
            }
          }
        ]
      },
      "UpdateUserRequest": {
        "type": "object",
        "description": "At least one field must be provided. confirmPassword is required when password is supplied.",
        "minProperties": 1,
        "additionalProperties": false,
        "properties": {
          "firstName": { "type": "string", "minLength": 3 },
          "lastName": { "type": "string", "minLength": 3 },
          "email": { "type": "string", "format": "email" },
          "phoneCountryCode": { "type": "string", "pattern": "^\\+[0-9]{1,4}$" },
          "phone": {
            "type": "string",
            "pattern": "^[0-9]{4,15}$",
            "description": "Digits only; spaces are removed before validation."
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "description": "Must contain uppercase, lowercase, numeric, and special characters."
          },
          "confirmPassword": { "type": "string", "minLength": 8 },
          "idNumber": { "type": "string", "minLength": 1 },
          "csdNumber": { "type": "string", "description": "Optional Central Securities Depository number" },
          "passportPhoto": {
            "type": "string",
            "description": "Secure URL returned by the Cloudinary upload endpoint for the user's passport photo"
          },
          "idDocument": {
            "type": "string",
            "description": "Secure URL returned by the Cloudinary upload endpoint for the user's identification document"
          },
          "dateOfBirth": { "type": "string", "format": "date" },
          "gender": {
            "type": "string",
            "enum": ["male", "female"]
          },
          "country": { "type": "string", "minLength": 1 },
          "city": { "type": "string", "minLength": 1 },
          "occupation": { "type": "string", "minLength": 1 },
          "investmentExperience": { "type": "string", "minLength": 1 },
          "notificationPreferences": {
            "type": "object",
            "additionalProperties": true
          },
          "role": { "$ref": "#/components/schemas/UserRole" },
          "isVerified": { "type": "boolean" }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "phoneCountryCode": { "type": "string" },
          "phone": { "type": "string" },
          "idNumber": { "type": "string" },
          "csdNumber": { "type": "string", "nullable": true },
          "passportPhoto": {
            "type": "string",
            "description": "Secure Cloudinary URL for the passport photo"
          },
          "idDocument": {
            "type": "string",
            "description": "Secure Cloudinary URL for the identification document"
          },
          "dateOfBirth": { "type": "string", "format": "date-time" },
          "gender": {
            "type": "string",
            "enum": ["male", "female"],
            "nullable": true
          },
          "country": { "type": "string" },
          "city": { "type": "string" },
          "occupation": { "type": "string" },
          "investmentExperience": { "type": "string" },
          "notificationPreferences": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "role": { "$ref": "#/components/schemas/UserRole" },
          "isVerified": { "type": "boolean" },
          "createdBy": {
            "type": "object",
            "nullable": true,
            "properties": {
              "id": { "type": "string", "format": "uuid" },
              "firstName": { "type": "string" },
              "lastName": { "type": "string" },
              "email": { "type": "string", "format": "email" },
              "role": { "$ref": "#/components/schemas/UserRole" }
            }
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "phoneCountryCode",
          "phone",
          "idNumber",
          "dateOfBirth",
          "country",
          "city",
          "occupation",
          "investmentExperience",
          "passportPhoto",
          "idDocument",
          "role",
          "isVerified",
          "createdAt",
          "updatedAt"
        ]
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/User" }
        },
        "required": ["data"]
      },
      "UserListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/User" }
          }
        },
        "required": ["data"]
      },
...
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" }
        },
        "required": ["message"]
      },
...
    }
  }
}